Refactors package structure for financial data classes and updates imports
